feat: remove Excel export functionality from various controllers and services
