+++ src/modules/promotion/promotion.controller.ts
@@ -9,7 +9,6 @@ import {
 	Req,
 	Query,
 	BadRequestException,
-	Res,
 } from '@nestjs/common';
 import { PromotionService } from './promotion.service';
 import { CreatePromotionRequestDto } from './dto/createPromotion.request.dto';
@@ -22,7 +21,6 @@ import { RequestWithUser } from '@/types/request.type';
 import { ApiPaginationQuery } from '@/decorators/apiPaginationQuery.decorator';
 import { ApiOperation, ApiQuery } from '@nestjs/swagger';
 import { PaginateData, PaginateParams, SortOrder } from '@/types/common.type';
-import { Response } from 'express';
 
 @Controller('promotions')
 @UseGuards(JwtAccessTokenGuard, RolesGuard)
@@ -73,15 +71,6 @@ export class PromotionController {
 		}
 
 		return this.promotionService.getAllPromotions(user.role, user.id, query);
-	}
-
-	@Get('export-promotion-excel')
-	@Roles(UserRole.Admin)
-	@ApiOperation({
-		summary: 'Export promotions to Excel',
-	})
-	async exportPromotionsToExcel(@Res() res: Response) {
-		return this.promotionService.exportPromotionsToExcel(res);
 	}
 
 	@Get(':id')
